requirements.txt
asgiref==3.8.1
certifi==2025.7.9
charset-normalizer==3.4.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0